feat: Separate Function and page
기능과 페이지를 분리하여 관리하도록 구성했습니다.
