+++ pages/user/login.js
@@ -1,4 +1,4 @@
-import { login } from "../../source/api/certified/login_api"
+import { loginHandler } from "../../source/js/login";
 
 
 export function loginRender() {
@@ -29,17 +29,5 @@ export function loginRender() {
     </div>
   </section>
   `;
-
-  // 로그인 기능 
-
-  const loginForm = document.querySelector(".login_form")
-
-  loginForm.addEventListener("submit", (e) => {
-    e.preventDefault()
-    const body = {
-      email: e.target[0].value,
-      password: e.target[1].value
-    }
-    login("POST", body)
-  })
+  loginHandler()
 }
